Setup basic auto named commands
One slight pitfall. Because DefaultCommands are explicitly disabled in auto, we must do one of the following:
a) continuously run a position command, making it never end. In this case, it is much simpler, but we loose the ability to wait for X thing to reach target position before then moving on to Y (while still holding)
b) make a seperate hold posiiton command that simply uses velocity control to hold the height of the elevator. This command must be a ScheduleCommand to not block the advancement of the command groups. This works well, but requires proxying almost every command as any command not proxied requiring a subsystem the same as a proxied command will cancel the entire command composition, including the schedule command inside.

Which one is better? Do we need this nuanced ability?